% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKgrid}
\alias{UKgrid}
\title{The UK National Electricity Transmission System Dataset}
\format{Data frame with timestamp (half-hour intervals)}
\source{
The UK natioanl grid \href{https://www.nationalgrid.com/uk}{link}
}
\usage{
UKgrid
}
\description{
The demand for electricity in the UK since 2011
Units: Thousands of units
}
\examples{
data(UKgrid)
}
\keyword{datasets}
