% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{extract_grid}
\alias{extract_grid}
\title{Extracting and transforming data from the UKgrid Dataset}
\usage{
extract_grid(type = "xts", columns = "ND", start = NULL, end = NULL,
  aggregate = NULL, weekly_agg = "index")
}
\arguments{
\item{type}{A character, define the output type - c("xts", "zoo", "ts", "mts", "data.frame", "tbl", "data.table")}

\item{columns}{Selecting the columns names to be used from the UKgrid dataset,
can be either the numeric values of the columns index, or a string with the column names}

\item{start}{Defines the starting date and time of the data extractions,
could be either an integer with the year value (4-digits format) or Date/POSIXt obejct}

\item{end}{Defines the ending date and time of the data extractions,
could be either an integer with the year value (4-digits format) or Date/POSIXt obejct}

\item{aggregate}{A string, if not NULL (default) aggregate up the series.
possible aggregation options are c("hourly", "daily", "weekly", "monthly", "quarterly", "yearly")}

\item{weekly_agg}{A string, set the week of the year definition if the argument "aggregate"  is set to "weekly".
Possible options "index" (default), "week", "isoweek", or "epiweek"}
}
\description{
Extracting and transforming data from the UKgrid Dataset
}
\examples{
df <- extract_grid(type = "xts", columns = "ND", start = 2017)

}
